Audit sweep: clear stray components/common imports from viz compute dirs
- Move formatValueForTooltip (components/ChartTooltip/utils.ts) and the
  stacked tooltip row helpers (StackedDataTooltip/utils.ts) into
  echarts/tooltip/{format,utils}.ts - pure compute used by echarts option
  builders; UI consumers updated, no shims left behind.
- Move the plain memoize function from common/hooks/use-memoized-callback
  into metabase/utils/memoize.ts (its natural home next to memoizeClass);
  the hook and all consumers import it from there.
