+++ frontend/src/metabase/utils/memoize.ts
@@ -1,5 +1,52 @@
 type Constructor<T> = new (...args: any[]) => T;
 
+type MapLike<K, V> = Map<K, V> | WeakMap<WeakKey, V>;
+
+function getMapLikeWithFallback<K, V>(
+  map: MapLike<K, V>,
+  key: K,
+  fallback: () => V,
+): V {
+  // Unjustified type cast. FIXME
+  if ("has" in map && map.has(key as any)) {
+    // Unjustified type cast. FIXME
+    return map.get(key as any)!;
+  } else {
+    const value = fallback();
+    // Unjustified type cast. FIXME
+    map.set(key as any, value);
+    return value;
+  }
+}
+
+type CacheValue = unknown;
+type CacheMap = Map<unknown, CacheValue | CacheMap>;
+
+const rootCache = new WeakMap<object, CacheMap>();
+const createCacheMap = (): CacheMap => new Map();
+
+export function memoize<T extends (...args: any[]) => any>(fn: T): T {
+  const memoizedFn = (...args: Parameters<T>): ReturnType<T> => {
+    const [first, ...rest] = [fn, args.length, ...args];
+    const lastKey = rest.pop()!;
+
+    let currentMap = getMapLikeWithFallback(rootCache, first, createCacheMap);
+    for (const key of rest) {
+      const nextMap = getMapLikeWithFallback(currentMap, key, createCacheMap);
+      if (nextMap instanceof Map) {
+        currentMap = nextMap;
+      } else {
+        throw new Error("Invalid cache structure");
+      }
+    }
+
+    return getMapLikeWithFallback(currentMap, lastKey, () => fn(...args));
+  };
+
+  // Unjustified type cast. FIXME
+  return memoizedFn as T;
+}
+
 function getWithFallback(
   map: Map<string, any>,
   key: string,
